requirements.txt
Flask
requests
openai
gunicorn
newspaper3k
beautifulsoup4
lxml
